# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sqlx-todos",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/stories": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Get a page of stories",
        "operationId": "get_stories",
        "parameters": [
          {
            "name": "page_size",
            "in": "query",
            "description": "The number of stories per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 1000,
              "minimum": 10
            }
          },
          {
            "name": "page_token",
            "in": "query",
            "description": "The next page cursor (next_page from response)",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of stories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stories"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Story"
        ],
        "summary": "Create a new story",
        "operationId": "create_story",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The story was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                }
              }
            }
          },
          "400": {
            "description": "The request body was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/stories/{story_id}": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Get a story",
        "operationId": "get_story",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The story",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                }
              }
            }
          },
          "404": {
            "description": "The story was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Story"
        ],
        "summary": "Delete a story",
        "operationId": "delete_story",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The story was deleted"
          },
          "404": {
            "description": "The story was not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Story"
        ],
        "summary": "Update a story",
        "operationId": "update_story",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The story was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Story"
                }
              }
            }
          },
          "400": {
            "description": "The request body was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "404": {
            "description": "The story not not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
...
    "/stories/{story_id}/tasks": {
      "get": {
        "tags": [
          "Story"
        ],
        "summary": "Get tasks for a story",
        "operationId": "get_tasks",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "description": "The story id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The task status filter",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tasks for the story",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "post": {
        "tags": [
          "Task"
        ],
        "summary": "Create a task",
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The task was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "The requesst body was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "Task"
        ],
        "summary": "Get a task",
        "operationId": "get_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "The task was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Task"
        ],
        "summary": "Delete a task",
        "operationId": "delete_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The task was deleted"
          },
          "404": {
            "description": "The task was not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Task"
        ],
        "summary": "Update a task",
        "operationId": "update_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "The task id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The task was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "The request body was invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          },
          "404": {
            "description": "The task was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errors"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTaskRequest": {
        "type": "object",
        "description": "The POST body for creating tasks",
        "required": [
          "name",
          "story_id"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "story_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Errors": {
        "type": "object",
        "description": "The type sent as an error response to the client.",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "incomplete",
          "complete"
        ]
      },
      "Stories": {
        "type": "object",
        "description": "A page of stories",
        "required": [
          "stories"
        ],
        "properties": {
          "next_page": {
            "type": "string",
            "nullable": true
          },
          "stories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Story"
            }
          }
        }
      },
      "Story": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StoryRequest": {
        "type": "object",
        "description": "The request body for creating or updating stories",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "required": [
          "id",
          "story_id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "story_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "description": "The PATCH body for updating tasks",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  },
...
}